mainGUI/widgetにウィジェット類を格納
